requirements.txt
amqp==5.1.1
asgiref==3.5.2
attrs==22.2.0
billiard==3.6.4.0
celery==5.2.7
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
Django==4.1.2
django-braces==1.15.0
django-celery-beat==2.4.0
django-ckeditor==6.5.1
django-colorfield==0.7.2
django-filter==22.1
django-js-asset==2.0.0
django-model-utils==4.3.1
django-star-ratings==0.9.2
django-timezone-field==5.0
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.25.1
idna==3.4
inflection==0.5.1
jsonschema==4.17.3
kavenegar==1.1.2
kombu==5.2.4
Pillow==9.3.0
polib==1.1.1
prompt-toolkit==3.0.36
psycopg2==2.9.5
PyJWT==2.6.0
pyrsistent==0.19.3
python-crontab==2.7.1
python-dateutil==2.8.2
python-decouple==3.8
pytz==2022.6
PyYAML==6.0
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
swapper==1.3.0
tzdata==2022.6
uritemplate==4.1.1
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
